Decompose action bar actions
