+++ graphcap_studio/src/features/gallery-viewer/components/actions/index.ts
@@ -1,0 +1,6 @@
+// SPDX-License-Identifier: Apache-2.0
+export * from "./AddToDatasetAction";
+export * from "./DeleteAction";
+export * from "./DownloadAction";
+export * from "./EditAction";
+
